destructure app component - add thememanager component
